refactor: graph
