+++ src/constants/index.ts
@@ -20,5 +20,5 @@ export const GRAPH_NAME_HEIGHT = 10
 
 export const GRAPH_PADDING_VERTICAL = 50
 export const GRAPH_PADDING_HORIZONTAL = 50
-export const GRAPH_MINIMUM_PADDING_VERTICAL = 5
-export const GRAPH_MINIMUM_PADDING_HORIZONTAL = 5
+export const GRAPH_WITHOUT_ROOT_PADDING_VERTICAL = 5
+export const GRAPH_WITHOUT_ROOT_PADDING_HORIZONTAL = 5
